backend/pyproject.toml
[project]
name = "rss-backend"
version = "0.1.0"
description = "RSS backend service"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.111",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "pydantic>=2.6",
    "httpx>=0.27",
    "feedparser>=6.0",
    "tenacity>=8.2",
    "apscheduler>=3.10",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "readability-lxml>=0.8",
    "lxml>=5.1",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "coverage>=7.4",
]
